src/requirements.txt
Keras>=2.1.1
opencv-python>=4.4.0.44
tensorflow-gpu>=1.15.0
torch>=1.4.0
torchvision>=0.5.0
matplotlib
tqdm
pip>=21.0.1
wheel>=0.36.2
setuptools>=41.2.0
numpy>=1.0.5
pillow>=5.1.0
scipy>=0.6.1
pandas>=0.2.3
openpyxl>=1.0.1
matplotlib>=2.3.4
facenet-pytorch>=2.5.2
seaborn
pandas
h5py==2.10.0
scikit-image>=0.18.2